Move ClassEntity to parser internal
