+++ core/src/test/java/com/nikodoko/javaimports/parser/ClassExtenderTest.java
@@ -3,7 +3,7 @@ package com.nikodoko.javaimports.parser;
 import static com.google.common.truth.Truth.assertThat;
 
 import com.google.common.collect.ImmutableSet;
-import com.nikodoko.javaimports.parser.entities.ClassEntity;
+import com.nikodoko.javaimports.parser.internal.ClassEntity;
 import com.nikodoko.javaimports.parser.internal.ClassSelectors;
 import org.junit.jupiter.api.Test;
